Extract notify and utils to the multirust_utils crate
