+++ Cargo.toml
@@ -17,9 +17,9 @@ license = "MIT OR Apache-2.0"
 
 [dependencies]
 multirust-dist = { version = "0.0.5", path = "src/multirust-dist" }
+multirust-utils = { version = "0.0.5", path = "src/multirust-utils" }
 clap = "1.4.5"
 regex = "0.1.41"
-rand = "0.3.11"
 openssl = "0.7.2"
 hyper = "0.7.0"
 term = "0.2.11"
