+++ src/multirust-dist/Cargo.toml
@@ -13,7 +13,6 @@ license = "MIT OR Apache-2.0"
 
 [dependencies]
 regex = "0.1.41"
-rand = "0.3.11"
 hyper = "0.7.0"
 openssl = "0.7.2"
 itertools = "0.4.1"
@@ -24,6 +23,7 @@ tempdir = "0.3.4"
 walkdir = "0.1.5"
 toml = "0.1.27"
 multirust-mock = { path = "../multirust-mock" }
+multirust-utils = { path = "../multirust-utils" }
 
 [target.x86_64-pc-windows-gnu.dependencies]
 winapi = "0.2.4"
