+++ src/multirust-dist/tests/transactions.rs
@@ -1,10 +1,13 @@
 extern crate multirust_dist;
+extern crate multirust_utils;
 extern crate tempdir;
 
 use multirust_dist::NotifyHandler;
 use multirust_dist::prefix::InstallPrefix;
 use multirust_dist::component::Transaction;
-use multirust_dist::{temp, utils};
+use multirust_dist::temp;
+use multirust_utils::utils;
+use multirust_utils::raw as utils_raw;
 use multirust_dist::Error;
 use tempdir::TempDir;
 use std::fs;
@@ -30,7 +33,7 @@ fn add_file() {
     tx.commit();
     drop(file);
 
-    assert_eq!(utils::raw::read_file(&prefix.path().join("foo/bar")).unwrap(),
+    assert_eq!(utils_raw::read_file(&prefix.path().join("foo/bar")).unwrap(),
                "test");
 }
 
@@ -403,7 +406,7 @@ fn write_file() {
 
     let path = prefix.path().join("foo/bar");
     assert!(utils::is_file(&path));
-    let file_content = utils::raw::read_file(&path).unwrap();
+    let file_content = utils_raw::read_file(&path).unwrap();
     assert_eq!(content, file_content);
 }
 
@@ -441,7 +444,7 @@ fn write_file_that_exists() {
     let mut tx = Transaction::new(prefix.clone(), &tmpcfg, notify);
 
     let content = "hi".to_string();
-    utils::raw::write_file(&prefix.path().join("a"), &content).unwrap();
+    utils_raw::write_file(&prefix.path().join("a"), &content).unwrap();
     let err = tx.write_file("c", PathBuf::from("a"), content.clone()).unwrap_err();
 
     match err {
@@ -490,11 +493,11 @@ fn modify_file_that_exists() {
     let mut tx = Transaction::new(prefix.clone(), &tmpcfg, notify);
 
     let ref path = prefix.path().join("foo");
-    utils::raw::write_file(path, "wow").unwrap();
+    utils_raw::write_file(path, "wow").unwrap();
     tx.modify_file(PathBuf::from("foo")).unwrap();
     tx.commit();
 
-    assert_eq!(utils::raw::read_file(path).unwrap(), "wow");
+    assert_eq!(utils_raw::read_file(path).unwrap(), "wow");
 }
 
 #[test]
@@ -530,12 +533,12 @@ fn modify_file_that_exists_then_rollback() {
     let mut tx = Transaction::new(prefix.clone(), &tmpcfg, notify);
 
     let ref path = prefix.path().join("foo");
-    utils::raw::write_file(path, "wow").unwrap();
+    utils_raw::write_file(path, "wow").unwrap();
     tx.modify_file(PathBuf::from("foo")).unwrap();
-    utils::raw::write_file(path, "eww").unwrap();
+    utils_raw::write_file(path, "eww").unwrap();
     drop(tx);
 
-    assert_eq!(utils::raw::read_file(path).unwrap(), "wow");
+    assert_eq!(utils_raw::read_file(path).unwrap(), "wow");
 }
 
 // This is testing that the backup scheme is smart enough not
@@ -554,14 +557,14 @@ fn modify_twice_then_rollback() {
     let mut tx = Transaction::new(prefix.clone(), &tmpcfg, notify);
 
     let ref path = prefix.path().join("foo");
-    utils::raw::write_file(path, "wow").unwrap();
+    utils_raw::write_file(path, "wow").unwrap();
     tx.modify_file(PathBuf::from("foo")).unwrap();
-    utils::raw::write_file(path, "eww").unwrap();
+    utils_raw::write_file(path, "eww").unwrap();
     tx.modify_file(PathBuf::from("foo")).unwrap();
-    utils::raw::write_file(path, "ewww").unwrap();
+    utils_raw::write_file(path, "ewww").unwrap();
     drop(tx);
 
-    assert_eq!(utils::raw::read_file(path).unwrap(), "wow");
+    assert_eq!(utils_raw::read_file(path).unwrap(), "wow");
 }
 
 fn do_multiple_op_transaction(rollback: bool) {
@@ -601,29 +604,29 @@ fn do_multiple_op_transaction(rollback: bool) {
 
     let ref srcpath1 = srcdir.path().join(&relpath1);
     fs::create_dir_all(srcpath1.parent().unwrap()).unwrap();
-    utils::raw::write_file(srcpath1, "").unwrap();
+    utils_raw::write_file(srcpath1, "").unwrap();
     tx.copy_file("", relpath1, srcpath1).unwrap();
 
     let ref srcpath2 = srcdir.path().join(&relpath2);
-    utils::raw::write_file(srcpath2, "").unwrap();
+    utils_raw::write_file(srcpath2, "").unwrap();
     tx.copy_file("", relpath2, srcpath2).unwrap();
 
     let ref srcpath4 = srcdir.path().join(&relpath4);
     fs::create_dir_all(srcpath4.parent().unwrap()).unwrap();
-    utils::raw::write_file(srcpath4, "").unwrap();
+    utils_raw::write_file(srcpath4, "").unwrap();
     tx.copy_dir("", PathBuf::from("doc"), &srcdir.path().join("doc")).unwrap();
 
     tx.modify_file(relpath5).unwrap();
-    utils::raw::write_file(path5, "").unwrap();
+    utils_raw::write_file(path5, "").unwrap();
 
     tx.write_file("", relpath6, "".to_string()).unwrap();
 
     fs::create_dir_all(path7.parent().unwrap()).unwrap();
-    utils::raw::write_file(path7, "").unwrap();
+    utils_raw::write_file(path7, "").unwrap();
     tx.remove_file("", relpath7).unwrap();
 
     fs::create_dir_all(path8.parent().unwrap()).unwrap();
-    utils::raw::write_file(path8, "").unwrap();
+    utils_raw::write_file(path8, "").unwrap();
     tx.remove_dir("", PathBuf::from("olddoc")).unwrap();
 
     if !rollback {
