+++ src/multirust/install.rs
@@ -2,8 +2,8 @@
 //! toolchains
 
 use multirust_dist::{NotifyHandler, Notification};
-use multirust_dist::utils;
 use multirust_dist::prefix::InstallPrefix;
+use multirust_utils::utils;
 use multirust_dist::temp;
 use multirust_dist::dist;
 use multirust_dist::component::{Components, TarGzPackage, Transaction, Package};
