+++ src/multirust/lib.rs
@@ -1,6 +1,6 @@
-#[macro_use]
 extern crate multirust_dist;
-extern crate rand;
+#[macro_use]
+extern crate multirust_utils;
 extern crate hyper;
 extern crate regex;
 extern crate itertools;
@@ -9,7 +9,7 @@ pub use errors::*;
 pub use config::*;
 pub use toolchain::*;
 pub use override_db::*;
-pub use multirust_dist::{utils, notify};
+pub use multirust_utils::{utils, notify};
 
 mod errors;
 mod override_db;
