+++ tests/cli-v1.rs
@@ -2,6 +2,7 @@
 //! derived from multirust/test-v2.sh
 
 extern crate multirust_dist;
+extern crate multirust_utils;
 extern crate multirust_mock;
 extern crate tempdir;
 
@@ -11,7 +12,7 @@ use multirust_mock::clitools::{self, Config, Scenario,
                                expect_ok, expect_stdout_ok, expect_err,
                                expect_stderr_ok, set_current_dist_date,
                                change_dir, run};
-use multirust_dist::utils;
+use multirust_utils::utils;
 
 pub fn setup(f: &Fn(&Config)) {
     clitools::setup(Scenario::SimpleV1, f);
@@ -164,11 +165,11 @@ fn remove_override_toolchain_error_handling() {
 fn bad_sha_on_manifest() {
     setup(&|config| {
         let sha_file = config.distdir.path().join("dist/channel-rust-nightly.sha256");
-        let sha_str = utils::raw::read_file(&sha_file).unwrap();
+        let sha_str = multirust_utils::raw::read_file(&sha_file).unwrap();
         let mut sha_bytes = sha_str.into_bytes();
         &mut sha_bytes[..10].clone_from_slice(b"aaaaaaaaaa");
         let sha_str = String::from_utf8(sha_bytes).unwrap();
-        utils::raw::write_file(&sha_file, &sha_str).unwrap();
+        multirust_utils::raw::write_file(&sha_file, &sha_str).unwrap();
         expect_err(config, &["multirust", "default", "nightly"],
                    "checksum failed");
     });
@@ -181,7 +182,7 @@ fn bad_sha_on_installer() {
         for file in fs::read_dir(&dir).unwrap() {
             let file = file.unwrap();
             if file.path().to_string_lossy().ends_with(".tar.gz") {
-                utils::raw::write_file(&file.path(), "xxx").unwrap();
+                multirust_utils::raw::write_file(&file.path(), "xxx").unwrap();
             }
         }
         expect_err(config, &["multirust", "default", "nightly"],
